# 8284075: Optimize paired floorDiv and floorMod calculations in java.time