--- a/src/java.base/share/classes/java/time/Duration.java
+++ b/src/java.base/share/classes/java/time/Duration.java
@@ -246,9 +246,9 @@ public final class Duration
      * @throws ArithmeticException if the adjustment causes the seconds to exceed the capacity of {@code Duration}
      */
     public static Duration ofSeconds(long seconds, long nanoAdjustment) {
-        long secs = Math.addExact(seconds, Math.floorDiv(nanoAdjustment, NANOS_PER_SECOND));
-        int nos = (int) Math.floorMod(nanoAdjustment, NANOS_PER_SECOND);
-        return create(secs, nos);
+        long secs = Math.floorDiv(nanoAdjustment, NANOS_PER_SECOND);
+        int nos = (int) (nanoAdjustment - secs * NANOS_PER_SECOND);
+        return create(Math.addExact(seconds, secs), nos);
     }
 
     //-----------------------------------------------------------------------