A common arithmetic pattern in java.time is to calculate both `Math.floorDiv` and `floorMod` for the same arguments. In such cases `floorMod` can be calculated more efficiently using the result of `floorDiv`.

From the javadoc for `floorMod`, the relationship is such that:

```java
floorDiv(x, y) * y + floorMod(x, y) == x
```
or:
```java
floorMod(x, y) == x - floorDiv(x, y) * y;
```

Simplifying like so bring a 15% or so speed-up on `GetYearBench`, which benefits from the changes `Instant::ofEpochMilli` and `LocalDateTime::ofEpochSecond`:

```
Benchmark                                                   Mode  Cnt       Score       Error   Units
GetYearBench.getYearFromMillisZoneOffset                   thrpt    5      34.522 ±     0.054  ops/ms
GetYearBench.getYearFromMillisZoneRegion                   thrpt    5       9.984 ±     0.151  ops/ms
GetYearBench.getYearFromMillisZoneRegionNormalized         thrpt    5      34.844 ±     0.281  ops/ms
GetYearBench.getYearFromMillisZoneRegionUTC                thrpt    5      34.517 ±     0.007  ops/ms
```

```
Benchmark                                                   Mode  Cnt       Score       Error   Units
GetYearBench.getYearFromMillisZoneOffset                   thrpt    5      39.709 ±     0.048  ops/ms
GetYearBench.getYearFromMillisZoneRegion                   thrpt    5      11.703 ±     0.510  ops/ms
GetYearBench.getYearFromMillisZoneRegionNormalized         thrpt    5      39.612 ±     0.023  ops/ms
GetYearBench.getYearFromMillisZoneRegionUTC                thrpt    5      39.511 ±     0.050  ops/ms
```

Testing: java.time tests locally, tier1+2 ongoing

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [ ] Change must be properly reviewed

### Issue
 * [JDK-8284075](https://bugs.openjdk.java.net/browse/JDK-8284075): Optimize paired floorDiv and floorMod calculations in java.time


### Reviewers
 * [Stephen Colebourne](https://openjdk.java.net/census#scolebourne) (@jodastephen - Author) ⚠️ Review applies to 37b33695e529b080faa75c2a0b87dd1829f8e5d8


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.java.net/jdk pull/8056/head:pull/8056` \
`$ git checkout pull/8056`

Update a local copy of the PR: \
`$ git checkout pull/8056` \
`$ git pull https://git.openjdk.java.net/jdk pull/8056/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 8056`

View PR using the GUI difftool: \
`$ git pr show -t 8056`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.java.net/jdk/pull/8056.diff">https://git.openjdk.java.net/jdk/pull/8056.diff</a>

</details>