@cl4es 2022-03-31
Hi @merykitty - thanks for taking a look. I tried your suggested ideas but haven't been able to make operation pair any faster. I suspect we would need a bit more JIT magic than making them look similar.