@merykitty 2022-03-31
@cl4es  I think I know why, in these cases, the divisors are constants so `a % b` is transformed into `a - b * (a / b)` anyway (and the division is transformed further). So my proposal is more or less the same thing as the operations today already. Thanks.