@theRealAph 2022-03-31
I'm not seeing a big difference, at least on AArch64. There is no divide instruction anywhere in the loop of `GetYearBench_getYearFromMillisZoneXXX_jmhTest`. To perform the division and the mod, we do a `mul` and a `mulh`, and a couple of shifts. The multiplication needed by the `%` and the `/` is only done once.

Before

```
Benchmark                                            Mode  Cnt   Score   Error   Units
GetYearBench.getYearFromMillisZoneOffset            thrpt    5  82.795 ± 0.038  ops/ms
GetYearBench.getYearFromMillisZoneRegion            thrpt    5  23.354 ± 0.030  ops/ms
GetYearBench.getYearFromMillisZoneRegionNormalized  thrpt    5  82.946 ± 0.043  ops/ms
GetYearBench.getYearFromMillisZoneRegionUTC         thrpt    5  82.263 ± 0.029  ops/ms
```
After

```
Benchmark                                            Mode  Cnt   Score   Error   Units
GetYearBench.getYearFromMillisZoneOffset            thrpt    5  84.649 ± 0.028  ops/ms
GetYearBench.getYearFromMillisZoneRegion            thrpt    5  24.183 ± 0.031  ops/ms
GetYearBench.getYearFromMillisZoneRegionNormalized  thrpt    5  84.220 ± 0.044  ops/ms
GetYearBench.getYearFromMillisZoneRegionUTC         thrpt    5  83.794 ± 0.975  ops/ms
```

I think this is almost in the noise, really