@cl4es 2022-03-31
@theRealAph the numbers I got was from AArch64 so I went back and double-checked. I think you're right and that I had merged in #8039 in my experiment branch but not my baseline. D'oh!

Baseline:
```
Benchmark                                            Mode  Cnt   Score   Error   Units
GetYearBench.getYearFromMillisZoneOffset            thrpt    5  39.719 ? 0.008  ops/ms
GetYearBench.getYearFromMillisZoneRegion            thrpt    5  11.658 ? 0.075  ops/ms
GetYearBench.getYearFromMillisZoneRegionNormalized  thrpt    5  39.616 ? 0.026  ops/ms
GetYearBench.getYearFromMillisZoneRegionUTC         thrpt    5  39.505 ? 0.142  ops/ms
```
Experiment:
```
Benchmark                                            Mode  Cnt   Score   Error   Units
GetYearBench.getYearFromMillisZoneOffset            thrpt    5  39.725 ? 0.050  ops/ms
GetYearBench.getYearFromMillisZoneRegion            thrpt    5  11.806 ? 0.017  ops/ms
GetYearBench.getYearFromMillisZoneRegionNormalized  thrpt    5  39.692 ? 0.265  ops/ms
GetYearBench.getYearFromMillisZoneRegionUTC         thrpt    5  39.661 ? 0.093  ops/ms
```

I'll withdraw this PR.